Restore OMX status panes as durable surfaces
The status pane is not an agent session, so app relaunch was rebuilding its saved leaf as a generic terminal while only the main Codex pane resumed. The restore path now persists a narrowly recognized OMX status tmux start command and relaunches that pane through a fresh session launcher script, while the existing split sizing work stays at pane creation time.

The CI guard also required extracting oversized Swift helpers into focused files instead of expanding file budgets.

Constraint: Build only through ./scripts/reload.sh --tag issue-3513-omx-hud-bottom-pane --launch; never direct xcodebuild
Constraint: Swift file length budget blocks PR checks
Rejected: Treat every tmux_start_command as restorable | would rerun arbitrary user split commands after app restart
Rejected: Revert Bonsplit initial divider support | would restore the first-render 50/50 status jump
Confidence: high
Scope-risk: moderate
Directive: OMX status startup belongs to session restoration metadata; do not reintroduce post-restore pane repair or generic tmux command replay
Tested: git diff --check
Tested: python3 scripts/swift_file_length_budget.py --repo-root .
Tested: plutil -lint GhosttyTabs.xcodeproj/project.pbxproj
Tested: ./scripts/reload.sh --tag issue-3513-omx-hud-bottom-pane --launch
Not-tested: Local test suite per repo policy
