+++ cmuxTests/WorkspaceUnitTests.swift
@@ -14,15 +14,6 @@ import Combine
 @testable import cmux
 #endif
 
-private func workspaceSplitNodes(in node: ExternalTreeNode) -> [ExternalSplitNode] {
-    switch node {
-    case .pane:
-        return []
-    case .split(let split):
-        return [split] + workspaceSplitNodes(in: split.first) + workspaceSplitNodes(in: split.second)
-    }
-}
-
 @MainActor
 func makeTemporaryBrowserProfile(named prefix: String) throws -> BrowserProfileDefinition {
     try XCTUnwrap(
@@ -3053,83 +3044,6 @@ final class WorkspaceSplitWorkingDirectoryTests: XCTestCase {
             requestedDirectory,
             "Expected split to inherit the source terminal's requested cwd when no reported cwd exists yet"
         )
-    }
-
-    func testTabManagerSplitCarriesRequestedWorkingDirectoryAndStartupCommand() {
-        let manager = TabManager()
-        guard let workspace = manager.selectedWorkspace,
-              let sourcePanelId = workspace.focusedPanelId else {
-            XCTFail("Expected selected workspace with a focused terminal")
-            return
-        }
-
-        let requestedDirectory = "/tmp/cmux-split-startup-\(UUID().uuidString)"
-        let startupCommand = "/tmp/cmux-tmux-command-\(UUID().uuidString).sh"
-        let tmuxStartCommand = "node /opt/oh-my-codex/dist/omx.js hud --watch"
-        let initialDividerPosition = 0.875
-        guard let splitPanelId = manager.newSplit(
-            tabId: workspace.id,
-            surfaceId: sourcePanelId,
-            direction: .down,
-            focus: false,
-            workingDirectory: requestedDirectory,
-            initialCommand: startupCommand,
-            tmuxStartCommand: tmuxStartCommand,
-            initialDividerPosition: initialDividerPosition
-        ) else {
-            XCTFail("Expected split terminal panel to be created")
-            return
-        }
-
-        let splitPanel = workspace.terminalPanel(for: splitPanelId)
-        XCTAssertEqual(splitPanel?.requestedWorkingDirectory, requestedDirectory)
-        XCTAssertEqual(
-            splitPanel?.surface.debugInitialCommand(),
-            startupCommand,
-            "Programmatic tmux-compatible splits must launch their command as the pane process"
-        )
-        XCTAssertEqual(
-            splitPanel?.surface.debugTmuxStartCommand(),
-            tmuxStartCommand,
-            "Programmatic tmux-compatible splits must preserve the original tmux command for pane format queries"
-        )
-        guard let split = workspaceSplitNodes(in: workspace.bonsplitController.treeSnapshot()).first else {
-            XCTFail("Expected split terminal panel to create a split node")
-            return
-        }
-        XCTAssertEqual(split.orientation, "vertical")
-        XCTAssertEqual(
-            split.dividerPosition,
-            initialDividerPosition,
-            accuracy: 0.000_1,
-            "Programmatic tmux-compatible splits should enter layout with their requested divider"
-        )
-    }
-
-    func testNewTerminalSurfaceCarriesRequestedWorkingDirectoryAndStartupCommand() {
-        let workspace = Workspace()
-        guard let paneId = workspace.bonsplitController.focusedPaneId else {
-            XCTFail("Expected focused pane in new workspace")
-            return
-        }
-
-        let requestedDirectory = "/tmp/cmux-surface-startup-\(UUID().uuidString)"
-        let startupCommand = "/tmp/cmux-surface-command-\(UUID().uuidString).sh"
-        let tmuxStartCommand = "node /opt/oh-my-codex/dist/omx.js hud --watch"
-        guard let surface = workspace.newTerminalSurface(
-            inPane: paneId,
-            focus: false,
-            workingDirectory: requestedDirectory,
-            initialCommand: startupCommand,
-            tmuxStartCommand: tmuxStartCommand
-        ) else {
-            XCTFail("Expected terminal surface to be created")
-            return
-        }
-
-        XCTAssertEqual(surface.requestedWorkingDirectory, requestedDirectory)
-        XCTAssertEqual(surface.surface.debugInitialCommand(), startupCommand)
-        XCTAssertEqual(surface.surface.debugTmuxStartCommand(), tmuxStartCommand)
     }
 
     func testNewTerminalSplitSkipsFreedInheritedSurfacePointer() throws {
